requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
pytz==2023.3
gigachat>=0.1.17
